Changed name, telebot -> framebot
