+++ CMakeLists.txt
@@ -1,5 +1,5 @@
 cmake_minimum_required (VERSION 3.7)
-project(Telebot)
+project(FramebotFramework)
 
 add_subdirectory(lib)
 add_subdirectory(test)
