+++ include/framebot/format.h
@@ -1,7 +1,7 @@
 #ifndef FORMAT_H
 #define FORMAT_H
 
-#include <telebot/telebot.h>
+#include <framebot/framebot.h>
 
 char *format(char *format, ...);
 char *vsformat (char *formats, va_list params);
