+++ include/framebot/json.h
@@ -1,7 +1,7 @@
 #ifndef JSON_H
 #define JSON_H
 
-#include <telebot/telebot.h>
+#include <framebot/framebot.h>
 
 json_t * start_json(char * json);
 void error_parse(json_t * error);
