+++ include/framebot/memory.h
@@ -1,7 +1,7 @@
 #ifndef MEMORY_H
 #define MEMORY_H
 
-#include <telebot/telebot.h>
+#include <framebot/framebot.h>
 
 char *alloc_string(const char * str);
 char *realloc_string(char *base, char *str);
