+++ include/framebot/objects.h
