+++ lib/memory.c
@@ -1,4 +1,4 @@
-#include <telebot/memory.h>
+#include <framebot/memory.h>
 
 char * alloc_string(const char * str){
