+++ test/CMakeLists.txt
@@ -1,5 +1,5 @@
 cmake_minimum_required (VERSION 3.7)
-project(TelebotTest)
+project(framebotTest)
 
 FIND_PACKAGE( OpenMP REQUIRED)
 if(OPENMP_FOUND)
@@ -15,7 +15,7 @@ include_directories (${PROJECT_SOURCE_DIR}/../include)
 add_definitions("-g")
 
 add_executable(test_lib test_lib.c)
-add_executable(libtest test_telebot.c)
+add_executable(libtest test_framebot.c)
 add_executable(update test_update.c)
 add_executable(file test_file.c)
 add_executable(userprofilephotos test_userprofilephotos.c)
@@ -29,33 +29,33 @@ add_executable(sendmessage test_sendmessage.c)
 add_executable(all_api test_api.c)
 
 if (CMAKE_SYSTEM_NAME STREQUAL Linux)
-	target_link_libraries(test_lib jansson libcurl.so telebot)
-	target_link_libraries(libtest jansson libcurl.so telebot)
-	target_link_libraries(update jansson libcurl.so telebot)
-	target_link_libraries(file jansson libcurl.so telebot)
-	target_link_libraries(userprofilephotos jansson libcurl.so telebot)
-	target_link_libraries(sendphotos jansson libcurl.so telebot)
-	target_link_libraries(sendaudio jansson libcurl.so telebot)
-	target_link_libraries(senddocument jansson libcurl.so telebot)
-	target_link_libraries(sendvideo jansson libcurl.so telebot)
-	target_link_libraries(sendvoice jansson libcurl.so telebot)
-	target_link_libraries(sendvideonote jansson libcurl.so telebot)
-	target_link_libraries(sendmessage jansson libcurl.so telebot)
-	target_link_libraries(all_api jansson libcurl.so telebot)
+	target_link_libraries(test_lib jansson libcurl.so framebot)
+	target_link_libraries(libtest jansson libcurl.so framebot)
+	target_link_libraries(update jansson libcurl.so framebot)
+	target_link_libraries(file jansson libcurl.so framebot)
+	target_link_libraries(userprofilephotos jansson libcurl.so framebot)
+	target_link_libraries(sendphotos jansson libcurl.so framebot)
+	target_link_libraries(sendaudio jansson libcurl.so framebot)
+	target_link_libraries(senddocument jansson libcurl.so framebot)
+	target_link_libraries(sendvideo jansson libcurl.so framebot)
+	target_link_libraries(sendvoice jansson libcurl.so framebot)
+	target_link_libraries(sendvideonote jansson libcurl.so framebot)
+	target_link_libraries(sendmessage jansson libcurl.so framebot)
+	target_link_libraries(all_api jansson libcurl.so framebot)
 endif (CMAKE_SYSTEM_NAME STREQUAL Linux)
 
 if (WIN32)
-	target_link_libraries(test_lib jansson libcurl_a telebot)
-	target_link_libraries(libtest jansson libcurl_a telebot)
-	target_link_libraries(update jansson libcurl_a telebot)
-	target_link_libraries(file jansson libcurl_a telebot)
-	target_link_libraries(userprofilephotos jansson libcurl_a telebot)
-	target_link_libraries(sendphotos jansson libcurl_a telebot)
-	target_link_libraries(sendphotos jansson libcurl_a telebot)
-	target_link_libraries(sendaudio jansson libcurl_a telebot)
-	target_link_libraries(senddocument jansson libcurl_a telebot)
-	target_link_libraries(sendvideo jansson libcurl_a telebot)
-	target_link_libraries(sendvideonote jansson libcurl_a telebot)
-	target_link_libraries(sendmessage jansson libcurl.so telebot)
-	target_link_libraries(all_api jansson libcurl.so telebot)
+	target_link_libraries(test_lib jansson libcurl_a framebot)
+	target_link_libraries(libtest jansson libcurl_a framebot)
+	target_link_libraries(update jansson libcurl_a framebot)
+	target_link_libraries(file jansson libcurl_a framebot)
+	target_link_libraries(userprofilephotos jansson libcurl_a framebot)
+	target_link_libraries(sendphotos jansson libcurl_a framebot)
+	target_link_libraries(sendphotos jansson libcurl_a framebot)
+	target_link_libraries(sendaudio jansson libcurl_a framebot)
+	target_link_libraries(senddocument jansson libcurl_a framebot)
+	target_link_libraries(sendvideo jansson libcurl_a framebot)
+	target_link_libraries(sendvideonote jansson libcurl_a framebot)
+	target_link_libraries(sendmessage jansson libcurl.so framebot)
+	target_link_libraries(all_api jansson libcurl.so framebot)
 endif (WIN32)
