+++ test/test_sendmessage.c
@@ -1,14 +1,14 @@
-#include <telebot/telebot.h>
+#include <framebot/framebot.h>
 
 Bot * _bot;
 
 int main(int argc, char *argv[]){
-	telebot_init();
+	framebot_init();
 
 	if(argc != 2)
 		fprintf(stderr, "sendphoto <token>");
 
-    _bot = telebot(argv[1]);
+    _bot = framebot(argv[1]);
 
 /* Message * send_photo_chat(Bot * bot, long int chat_id, char * filename,
 			  char * caption, bool disable_notification,
